added custom hook and removed duplicate code... to an extent
