+++ GreatProject.UI/src/Organism/matrix.js
@@ -1,33 +1,26 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import MATRIX_FRAMES from '../data/matrix';
 import './homepage.css';
-
+import { UseDynamicTransitions } from '../services/hooks';
 
 const minimumDelay = 1;
 const minimumIncrement = 1;
+
 export default function Matrix() {
 
-    const [index, setIndex] = useState(1);
     const [delay, setDelay] = useState(minimumDelay * 500);
     const [increment, setIncrement] = useState(minimumIncrement);
+    const index = UseDynamicTransitions({ delay, increment, length: MATRIX_FRAMES.length })
+
 
     const updateMatrixFrames = event => {
         var userDefinedDelay = Number(event.target.value);
-        setDelay((userDefinedDelay < minimumDelay) ? (minimumDelay * 1000) : (userDefinedDelay * 1000));
+        setDelay((userDefinedDelay < minimumDelay) ? (minimumDelay * 500) : (userDefinedDelay * 1000));
     }
     const updateIncrement = event => {
         var userDefinedIncrement = Number(event.target.value);
         setIncrement((userDefinedIncrement < minimumIncrement) ? minimumIncrement : userDefinedIncrement);
     }
-    useEffect(() => {
-        const interval = setInterval(() => {
-            setIndex((storedIndex) => {
-                return (storedIndex + increment) % MATRIX_FRAMES.length
-            })
-        }, delay);
-
-        return (() => { clearInterval(interval) })
-    }, [delay, increment, index]);
 
     return (
         <div className="Matrix">
